#4842. Fixed tons of 64-bit portability warnings in tools. Note: this only affects in-game editor tools, and only in 64-bit mode.
Also fixed the following bug, which also raised warning:
  LoadCursor ( NULL, MAKEINTRESOURCE(IDC_ARROW) )
Here MAKEINTRESOURCE is excessive: it is already applied to all IDC_XXX macros.


git-svn-id: http://svn.thedarkmod.com/svn/darkmod_src/trunk@7497 49c82d7f-2e2a-0410-a16f-ae8f201b507f

+++ tools/radiant/MapInfo.cpp
@@ -80,7 +80,7 @@ BOOL CMapInfo::OnInitDialog()
 
   CMapStringToPtr mapEntity;
 
-  int nValue = 0;
+  size_t nValue = 0;
 	for (entity_t* pEntity=entities.next ; pEntity != &entities ; pEntity=pEntity->next)
 	{
     m_nTotalEntities++;
